# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant B2B SaaS platform API.",
        "title": "Sapanjai API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/rbac/assign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rbac"
                ],
                "summary": "Assign a role to a member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Active organization ID",
                        "name": "x-organization-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Assignment payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_rbac.AssignRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_module_rbac.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Missing x-organization-id header",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not a member of this organization",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "MEMBER_NOT_FOUND / ROLE_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/rbac/roles": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rbac"
                ],
                "summary": "List roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Active organization ID",
                        "name": "x-organization-id",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_module_rbac.RoleResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Missing x-organization-id header",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not a member of this organization",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rbac"
                ],
                "summary": "Create a role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Active organization ID",
                        "name": "x-organization-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Role payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_rbac.CreateRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_module_rbac.RoleRowResponse"
                        }
                    },
                    "400": {
                        "description": "Missing x-organization-id header",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not a member of this organization",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/rbac/roles/{roleId}/permissions": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rbac"
                ],
                "summary": "Replace a role's permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Active organization ID",
                        "name": "x-organization-id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role ID",
                        "name": "roleId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New permission set",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_module_rbac.UpdatePermissionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_module_rbac.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Missing x-organization-id header",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not a member of this organization",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "ROLE_NOT_FOUND",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_sapanjai_backend_internal_shared_httpx.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_module_rbac.AssignRoleRequest": {
            "type": "object",
            "required": [
                "roleId",
                "userId"
            ],
            "properties": {
                "roleId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "internal_module_rbac.CreateRoleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "minLength": 1
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "internal_module_rbac.PermissionResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "roleId": {
                    "type": "string"
                }
            }
        },
        "internal_module_rbac.RoleResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_module_rbac.PermissionResponse"
                    }
                }
            }
        },
        "internal_module_rbac.RoleRowResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "string"
                }
            }
        },
        "internal_module_rbac.SuccessResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_module_rbac.UpdatePermissionsRequest": {
            "type": "object",
            "properties": {
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer {token}\" — the JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}